Revert "adds check for is cell editable"
This reverts commit 00b6b63cfaec81abe4a28a2e5684d6587552a069.
