+++ webapp/packages/plugin-data-spreadsheet-new/src/DataGrid/TableDataContext.ts
@@ -1,6 +1,6 @@
 /*
  * CloudBeaver - Cloud Database Manager
- * Copyright (C) 2020-2026 DBeaver Corp and others
+ * Copyright (C) 2020-2025 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0.
  * you may not use this file except in compliance with the License.
@@ -54,7 +54,6 @@ export interface ITableData {
   isIndexColumn: (columnKey: IColumnInfo) => boolean;
   isIndexColumnInRange: (columnsRange: Array<IColumnInfo>) => boolean;
   isCellReadonly: (key: IGridDataKey) => boolean;
-  isCellEditable: (rowIndex: number, columnIndex: number) => boolean;
 }
 
 export function isColumnInfo(column: IColumnInfo | undefined): column is IColumnInfo {
